Move explicit mouse organism model to own folder
This starts the work of breaking the Organism Space into unit-specific
folders and models, but this is only a preliminary job, as the mouse
model is tied quite deeply with the Organism Space.
